ems_doctrine_prototype/requirements.txt
# EMS Doctrine Prototype - Phase 0 Dependencies

# Core data processing
numpy>=1.24.0
pandas>=2.0.0

# PDF and document processing
PyMuPDF>=1.23.0
pdfplumber>=0.9.0

# Natural Language Processing
spacy>=3.6.0
transformers>=4.30.0
torch>=2.0.0

# Entity recognition and text processing
nltk>=3.8
textacy>=0.13.0

# Graph and knowledge representation
networkx>=3.1
rdflib>=6.3.0
SPARQLWrapper>=2.0.0

# Database and storage
sqlalchemy>=2.0.0
sqlite3

# Web API and interface
flask>=2.3.0
flask-cors>=4.0.0
requests>=2.31.0

# Data visualization and analysis
matplotlib>=3.7.0
plotly>=5.15.0
streamlit>=1.24.0

# Development and testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0

# Configuration and utilities
python-dotenv>=1.0.0
click>=8.1.0
pyyaml>=6.0.0